chore: move server code into src folder
